Refactor ResolvedPipelineTask and remove redundant fields
Currently, PipelineRun resolution code can be difficult to understand. This commit removes some of the redundant fields in  ResolvedPipelineTask including the singular version of: TaskRun, TaskRunName, ObjectRun, ObjectRunName so that only the list of TaskRuns, TaskRunNames,  ObjectRuns, ObjectRunNames can be used  regardless if the PipelineTask is matrixed or not. This also streamlines the codebase by removing the switch statements used throughout. This partially addresses Issue #6628.